喵

+++ src/io/kurumi/ntt/cqhttp/model/LoginInfo.java
@@ -2,7 +2,7 @@ package io.kurumi.ntt.cqhttp.model;
 
 public class LoginInfo {
 
-	Integer user_id;
+	Long user_id;
 	String nickname;
 
 }
